Made endpoints compatible with each respective book and not just Animal Farm.

+++ Model/Character.cs
@@ -5,7 +5,7 @@ using System.Threading.Tasks;
 
 public class Character
 {
-    public int EndSid { get; set; }
+    public float EndSid { get; set; }
     public string Name { get; set; }
     public string Summary { get; set; }
 }
